+++ chaosreliably/controls/experiment.py
@@ -9,7 +9,20 @@ from logzero import logger
 
 from chaosreliably import RELIABLY_HOST, get_session
 
-__all__ = ["after_experiment_control"]
+__all__ = ["after_experiment_control", "before_experiment_control"]
+
+
+def before_experiment_control(
+    context: Experiment,
+    exp_id: str,
+    org_id: str,
+    state: Journal,
+    configuration: Configuration = None,
+    secrets: Secrets = None,
+    **kwargs: Any,
+) -> None:
+    get_span(org_id, exp_id)
+    set_plan_status(org_id, "running", None, configuration, secrets)
 
 
 def after_experiment_control(
@@ -21,12 +34,7 @@ def after_experiment_control(
     secrets: Secrets = None,
     **kwargs: Any,
 ) -> None:
-
-    tracer = opentracing.global_tracer()
-    scope = tracer.scope_manager.active
-    span = scope.span if scope else None
-    if span:
-        span.set_tag("reliably-control", "started")
+    span = get_span(org_id, exp_id)
 
     try:
         result = complete_run(
@@ -39,6 +47,9 @@ def after_experiment_control(
 
             exec_id = payload["id"]
 
+            if span:
+                span.set_baggage_item("reliably_execution_id", exec_id)
+
             host = secrets.get(
                 "host", os.getenv("RELIABLY_HOST", RELIABLY_HOST)
             )
@@ -47,12 +58,16 @@ def after_experiment_control(
             extension["execution_url"] = url
 
             add_runtime_extra(extension)
+            set_plan_status(org_id, "completed", None, configuration, secrets)
     except Exception as ex:
         logger.debug(
             f"An error occurred: {ex}, while running the after-experiment "
             "control, the execution won't be affected.",
             exc_info=True,
         )
+        if span:
+            span.set_baggage_item("reliably_error", str(ex))
+        set_plan_status(org_id, "error", str(ex), configuration, secrets)
     finally:
         if span:
             span.set_tag("reliably-control", "finished")
@@ -101,3 +116,39 @@ def add_runtime_extra(extension: Dict[str, Any]) -> None:
         extension["extra"] = json.loads(extra)
     except Exception:
         pass
+
+
+def set_plan_status(
+    org_id: str,
+    status: str,
+    message: Optional[str],
+    configuration: Configuration,
+    secrets: Secrets,
+) -> None:
+    plan_id = os.getenv("RELIABLY_PLAN_ID")
+    if not plan_id:
+        return None
+
+    with get_session(configuration, secrets) as session:
+        resp = session.put(
+            f"/{org_id}/plans/{plan_id}/status",
+            json={"status": status, "error": message},
+        )
+        logger.debug(f"Response from {resp.url}: {resp.status_code}")
+
+
+def get_span(org_id: str, experiment_id: str) -> Optional[opentracing.Span]:
+    tracer = opentracing.global_tracer()
+    scope = tracer.scope_manager.active
+    span = scope.span if scope else None
+    if not span:
+        return None
+
+    span.set_baggage_item("reliably_org_id", org_id)
+    span.set_baggage_item("reliably_experiment_id", experiment_id)
+
+    plan_id = os.getenv("RELIABLY_PLAN_ID")
+    if plan_id:
+        span.set_baggage_item("reliably_plan_id", plan_id)
+
+    return span